Changed the sysex buffer and append commands to uint16_t.
This removes the need to split short values up into 2 bytes when Firmata will not expect them as such. If Firmata expects a byte on the other end, it can reassemble them as bytes. If, in the case of I2C, is expects a short, it can reassemble them that way. This simplified a previous bug fix in TwoWire.cpp